requirements.txt
# Visual Grounding – Python Dependencies
# Install: pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tqdm
numpy
Pillow
matplotlib
